REGRESSION(300277@main): [macOS Debug] 2x in fast/selectors/style-invalidation-hover-change are flaky failure https://bugs.webkit.org/show_bug.cgi?id=299553 rdar://161356964
Reviewed by Abrar Rahman Protyasha and Aditya Keerthi.

In 300277@main, we introduced hover state updates when a layout update results in
the element underneath the mouse changing. Both failing tests work by moving
the mouse into a specific element, and then checking style changes to see if the
hover state updated correctly. However, the output of these tests is positioned
above the elements, causing the tested elements to shift downwards each time a
PASS or FAIL message is printed. This causes the mouse underneath the element to
change. This did not cause a test failure before 300277@main because we simply
did not update the hover state correctly after layout changes.

After 300277@main, the test was flakey rather than simply failing due to a
race condition. If the style was checked before the layout change updated
the hover state, the test would pass, and if it checked afterwards, it would
fail due to the element under the mouse changing.

To fix this, simply add the console ourselves to the end of the page rather
than allowing the imported script to add it to the beginning of the page.

* LayoutTests/fast/selectors/style-invalidation-hover-change-descendants.html:
* LayoutTests/fast/selectors/style-invalidation-hover-change-siblings.html:
* LayoutTests/platform/mac/TestExpectations:

Canonical link: https://commits.webkit.org/300733@main