[GFC] Start storing some computed size properties on PlacedGridItem https://bugs.webkit.org/show_bug.cgi?id=299812 rdar://161584151
Reviewed by Alan Baradlay.

We will need access to various bits of a grid item's computed style
during layout. For example, the preferred sizes are properties that can
significantly impact the sizing of a grid item and influence the rest of the
grid layout. One option is to put the pieces of information that we need
on PlacedGridItem so that they can be accessed during layout, which is
what this patch aims to do.

We do this as we construct the PlacedGridItem from its unplaced counterpart
inside of GridFormattingContext. We need to do this here since
GridFormattingContext will have enough information to map the various
properties to the inline and block dimensions as appropriate. Here we
just assume horizontal-tb with non-orthogonal content for
simplicity.

In this patch I add the preferred sizes, min/max sizes, along with
margins as part of ComputedSizes. I leave out border and padding for now
since those are just additional sizes which get resolved to their used
value as part of updating the formatting context geometries before
layout. If we need access to those values we may want to consider
accessing them in a slightly different way since they are slightly
different from these ComputedSizes.

Canonical link: https://commits.webkit.org/300734@main