Web Inspector: Elements: Adding an attribute via the context menu on a closing tag throws an exception https://bugs.webkit.org/show_bug.cgi?id=299802 rdar://161577627
Reviewed by Devin Rousso.

The context menu for the element closing tag in the DOM tree offers the same actions
as for the opening tag, but they operate in the context of the `WI.DOMTreeElement` of the closing tag.

Technically, Web Inspector should not offer the actions, but,
as a convenience, we can redirect them to the context of the opening tag.

This solves the Add > Attribute exception as well as some UI glitches with
Add > Previous Sibling and Add > Child where the position of the added node shifts
after committing the value.

* Source/WebInspectorUI/UserInterface/Views/DOMTreeElement.js:
(WI.DOMTreeElement.prototype.populateDOMNodeContextMenu):

Canonical link: https://commits.webkit.org/300752@main