[Skia] Build a SkTextBlob when recording DrawGlyphs display list item https://bugs.webkit.org/show_bug.cgi?id=299869
Reviewed by Nikolas Zimmermann.

GlyphDisplayListCache uses a display list to cache the shaping results
of a text run that is frequently painted, but we are still building a
SkTextBlob for the same shaping results every time the display list is
replayed, which is expensive. We can build the SkTextBlob once when
recording and save that in DrawGlyphs item instead of the list of glyphs
and advances.

Canonical link: https://commits.webkit.org/300818@main