REGRESSION (iOS 26): MediaRecorder produces corrupted video/audio files, resulting in some loss of content https://bugs.webkit.org/show_bug.cgi?id=299164 rdar://160982017
Reviewed by Youenn Fablet.

In typical usage, a MediaRecorder is configured to generate segments of a finite size.
However, in the provided reproduction steps, no size are provided and the
audio recordings will be as long as the duration between the recording starting
and the recording being stopped.
When writing the last segment, if the AVAssetWriterInput reported not to be
ready for more data, we would abort in the expectation that another fetchData
would be called that would flush the remaining samples. However, this last fetchData
will never occur.

The MediaRecorderPrivateWriterAVFObjC made the incorrect assumptions that
the AVAssetWriterInput would always be accepting more content as we only feed
it compressed, interleaved sample. This assumption turned out to be false.

When closing the recording, we now pass to the writer all the pending samples
and write them asynchronously using the appropriate AVAssetWriterInput's APIs
that ensure writes are delayed until ready.
We now call AVAssetWriterInput as per its documentation suggests

Added test.

* Source/WebCore/platform/mediarecorder/MediaRecorderPrivateWriter.cpp:
(WebCore::MediaRecorderPrivateWriter::close):
* Source/WebCore/platform/mediarecorder/MediaRecorderPrivateWriter.h:
* Source/WebCore/platform/mediarecorder/cocoa/MediaRecorderPrivateWriterAVFObjC.h:
* Source/WebCore/platform/mediarecorder/cocoa/MediaRecorderPrivateWriterAVFObjC.mm:
(WebCore::MediaRecorderPrivateWriterAVFObjC::MediaRecorderPrivateWriterAVFObjC):
(WebCore::MediaRecorderPrivateWriterAVFObjC::writeFrame):
(WebCore::MediaRecorderPrivateWriterAVFObjC::close):
* Source/WebCore/platform/mediarecorder/cocoa/MediaRecorderPrivateWriterWebM.cpp:
(WebCore::MediaRecorderPrivateWriterWebM::close):
* Source/WebCore/platform/mediarecorder/cocoa/MediaRecorderPrivateWriterWebM.h:

Canonical link: https://commits.webkit.org/300824@main