Implement EventCounts as maplike https://bugs.webkit.org/show_bug.cgi?id=299211 rdar://160968888
Reviewed by Youenn Fablet.

Implements the EventCounts interface as a maplike, which
provides methods such as .forEach(), keys(), etc.

EventCounts cannot use the previous maplike implementation
because it would get out of sync after the backing map is
initialized. To solve this, we use the trait
shouldAlwaysInitializeMapLike to decide if the backing map
should be reloaded on every access through the wrapper. The trait
simply checks for the presence of a marker
(shouldAlwaysInitializeMapLikeMarker).

This implementation isn't perfect: although the backing map
will be kept up-to-date when accessed through the wrapper,
it won't be updated if accessed through an iterator generated
from that map. This is documented by the failing test
event-timing/event-counts-maplike-iterators-always-updated.html.

Canonical link: https://commits.webkit.org/300830@main