Strip StoragePartition from cookies returned by WKHTTPCookieStore API rdar://147233516 https://bugs.webkit.org/show_bug.cgi?id=292351
Reviewed by Alex Christensen.

The WKHTTPCookieStore API returns a NSHTTPCookie. When opt-in partitioned
cookies is enabled, a partitioned cookie can be set in either of a first or
third party context. As such, when a first party cookie is "partitioned", its
NSHTTPCookie will contain the "StoragePartition" property.  This is internal
state that does not need to be exposed to an application that uses the
WKHTTPCookieStore API, and it can cause breakage in some situations.

This change clears the cookie property before the cookie is returned to the
application when the WKHTTPCookieStore API is used.

Adds a new API test.

Test: Tools/TestWebKitAPI/Tests/WebKitCocoa/WKHTTPCookieStore.mm

Test: Tools/TestWebKitAPI/Tests/WebKitCocoa/WKHTTPCookieStore.mm
* Source/WebKit/UIProcess/API/APIHTTPCookieStore.cpp:
(API::HTTPCookieStore::isOptInCookiePartitioningEnabled const):
* Source/WebKit/UIProcess/API/APIHTTPCookieStore.h:
* Source/WebKit/UIProcess/API/Cocoa/WKHTTPCookieStore.mm:
(clearCookiePartitionPropertyIfNeeded):
(-[WKHTTPCookieStore getAllCookies:]):
(-[WKHTTPCookieStore _getCookiesForURL:completionHandler:]):
* Tools/TestWebKitAPI/Tests/WebKitCocoa/WKHTTPCookieStore.mm:
(TEST(WKHTTPCookieStore, PartitionedCookieShouldNotHavePartitionProperty)):

Canonical link: https://commits.webkit.org/300838@main