VM::m_entryScopeServicesRawBits needs to be initialized. https://bugs.webkit.org/show_bug.cgi?id=299974 rdar://161754896
Reviewed by Yusuke Suzuki and Keith Miller.

This was an oversight from 300049@main.

It will be hard to test for this bug because it depends on what the uninitialized value turns out to be
at runtime.  A test for an up-coming patch for a Stop The World mechanism will stress this a lot more
and may reveal the bug then.  That test is how I found this bug in the first place.

* Source/JavaScriptCore/runtime/VM.h:

Canonical link: https://commits.webkit.org/300843@main