[DumpRenderTree] Clear BackForwardList after each test. https://bugs.webkit.org/show_bug.cgi?id=299900 rdar://128479544
Reviewed by Per Arne Vollan and Rupin Mittal.

In this test: imported/w3c/web-platform-tests/html/browsers/history/the-history-interface/001.html

Test starts passes if run the test locally alone, but starts failing after some iteration.
It seems the previous back forward list items are remaining after test completed and eventually
it filled up with entries. In the test, it is expected that the history size is increased after
navigation but not because of the capacity.

This patch clear the list after each test. From the perspective of test isolation, this must be
the correct behavior.

* LayoutTests/TestExpectations:
* Tools/DumpRenderTree/mac/DumpRenderTree.mm:
(runTest):

Canonical link: https://commits.webkit.org/300848@main