[visionOS] make docking works when AudioVideoRendererRemote is in use https://bugs.webkit.org/show_bug.cgi?id=299520 rdar://161320965
Reviewed by Andy Estes.

We make the VideoReceiverEndpointManager be aware of the existing AudioVideoRenderers.

When a new VideoTargetEndpoing is received, if the lookup for a MediaPlayer
matching the required identifier fails, we will look into the RemoteAudioVideoRendererProxyManager's AudioVideoRenderers instead.

We make the MediaPlayerPrivateRemote communicates to the RemoteAudioVideoRendererProxyManager
the HTMLMediaElementIdentifier and MediaPlayerIdentifier and store it alongside the RendererContext.

Manually tested on Vision Pro.

Canonical link: https://commits.webkit.org/300854@main