Assert the assumption that two VM fields are loadable as a pair https://bugs.webkit.org/show_bug.cgi?id=299933 rdar://161707462
Reviewed by Mark Lam and Justin Michaud.

Fields `VM::topCallFrame` and `VM::topEntryFrame` are loaded and stored as a pair by the
assembly code. This PR adds a static assert to verify that the fields are laid out
accordingly, and a comment to point out that some loads and stores of `VM::topEntryFrame`
are implicit in pair instructions operating on `VM::topCallFrame`. A similar comment is
added to `VMEntryRecord`, which already has a static assert for its pair of fields used
with pair instructions.

The patch does not touch executable code and is not directly testable.

Canonical link: https://commits.webkit.org/300898@main