[css-anchor-position-1] Adjust order of position option to try when a last successful position option is present rdar://161714637 https://bugs.webkit.org/show_bug.cgi?id=299940
Reviewed by Elika Etemad.

The way we change the position option order to make the last successful position option
the first option to try is wrong. Previously, we took the position option list (after
sorting by position-try-order), and "rotated" it so that the last successful option
is the first option. This means the order goes like this:
1) last successful position option
2) option following the last successful position option in the before-rotated list
3) option following 2)
4) ...

This seems to be wrong, as the order to make WPT tests passes is:
1) last successful position option first
2) original style (not using any options)
3) other options in the sorted position option list, in order, excluding the last successful option

* LayoutTests/imported/w3c/web-platform-tests/css/css-anchor-position/anchor-scroll-position-try-006-expected.txt:
* LayoutTests/imported/w3c/web-platform-tests/css/css-anchor-position/anchor-scroll-position-try-008-expected.txt:
* LayoutTests/imported/w3c/web-platform-tests/css/css-anchor-position/anchor-scroll-position-try-009-expected.txt:
* LayoutTests/imported/w3c/web-platform-tests/css/css-anchor-position/anchor-scroll-position-try-010-expected.txt:
* Source/WebCore/style/StyleTreeResolver.cpp:
(WebCore::Style::TreeResolver::sortPositionOptionsIfNeeded):

Canonical link: https://commits.webkit.org/300909@main