SpeculationRules: Test that inline rules after an external stylesheet work https://bugs.webkit.org/show_bug.cgi?id=300078
Reviewed by NOBODY (OOPS!).

Explanation of why this fixes the bug (OOPS!).
ohttps://bugs.webkit.org/show_bug.cgi?id=300039
https://bugs.webkit.org/show_bug.cgi?id=300039 was a result of an untested path: inline speculation rules script after an external stylesheet.
Scripts after a stylesheet are handled in executePendingScript, and that code wasn't ready for speculation rules.
As a result, we triggered an ASSERT that didn't take the speculation rules type into account, and weren't registering speculation rules in that case.

This PR adds the relevant tests to make sure this path is exercised.

* LayoutTests/imported/w3c/web-platform-tests/speculation-rules/prefetch/inline-speculation-rules-after-stylesheet.https-expected.txt: Added.
* LayoutTests/imported/w3c/web-platform-tests/speculation-rules/prefetch/inline-speculation-rules-after-stylesheet.https.html: Added.